requirements.txt
docker-py>=0.3.1
